+++ native_dpcpp/kmeans/GPU/kmeans.h
@@ -15,12 +15,11 @@ void runKmeans(cl::sycl::queue* q,
 	       Point* points, 
 	       Centroid* centroids,
 	       size_t NUMBER_OF_POINTS,
-	       int NUMBER_OF_CENTROIDS
+	       size_t NUMBER_OF_CENTROIDS
 	       );
 
-void printCentroids(
-		    Centroid* centroids,
-		    int NUMBER_OF_CENTROIDS
+void printCentroids(Centroid* centroids,
+		    size_t NUMBER_OF_CENTROIDS
 		    );
 
 #endif
